refactor(backend): migrar litros de Double a BigDecimal
Los litros entraban a los calculos de plata como Double. El precio
unitario ya era BigDecimal, pero eso no salvaba nada: el operando LLEGABA
contaminado. El punto flotante binario no representa exactamente la
mayoria de los decimales, asi que el error de representacion ya estaba en
el valor antes de convertirlo (ver docs/BACKEND-AUDIT.md, DB-04).

Peor era ConsumoCalculator, que acumulaba `litros += ...` como double
primitivo a lo largo de muchos intervalos ANTES de envolver el total en
BigDecimal: ahi el error se sumaba vuelta a vuelta. En un sistema que
existe para controlar consumo de combustible, los numeros SON el producto.

Ahora es BigDecimal de punta a punta: entra BigDecimal por
CreateTicketRequest y por el borde del OCR (asDecimal), se acumula con
BigDecimal.add y sale BigDecimal por los DTOs. No queda ningun salto por
double en el medio.

REQUIERE MIGRACION. V12 hace ALTER COLUMN litros TYPE numeric(10,2). Se
hace ahora a proposito: produccion tiene del orden de 3 tickets, asi que
reescribir la tabla es gratis. Con volumen real esto seria una ventana de
mantenimiento.

NO ROMPE EL CONTRATO CON mobile/. Se verifico: BigDecimal se serializa a
JSON como numero, y la app ya tipa litros como `number` y lo envia como
string en un campo multipart. Cero cambios en mobile, cero APK nuevo por
este motivo. La auditoria habia previsto lo contrario; la prediccion
confundia "cambia el tipo en Java" con "cambia el contrato HTTP".

Dos tests nuevos en TicketPersistenciaTest prueban contra Postgres real lo
que la migracion compra: round-trip exacto de 45.67 (con equals, para
exigir tambien la escala) y que 0,10 + 0,20 da exactamente 0,30.

Al migrar la suite apareció que BigDecimal.equals compara TAMBIEN la
escala: 50.0 no es igual a 50.00. Dos tests fallaron por eso y no por
logica; se pasaron a compareTo / isEqualByComparingTo, que es lo correcto
cuando lo que se afirma es el valor.

Suite 302 -> 304.

+++ backend/src/main/java/com/retrorental/backend/dto/request/CreateTicketRequest.java
@@ -17,9 +17,13 @@ import java.time.LocalDateTime;
 @OrigenCargaCoherente
 public class CreateTicketRequest implements CargaOrigen {
 
+    // BigDecimal desde el BORDE de entrada, no convertido mas adentro: si se
+    // recibiera como Double, el valor ya llegaria con error de representacion y
+    // convertirlo despues no lo recupera (ver docs/BACKEND-AUDIT.md, DB-04).
+    // Jackson parsea el numero del JSON directo a BigDecimal, exacto.
     @NotNull(message = "Los litros son obligatorios")
     @Positive(message = "Los litros deben ser mayores a cero")
-    private Double litros;
+    private BigDecimal litros;
 
     // Opcional: si no viene, el backend usa el momento actual. No puede ser futura.
     @PastOrPresent(message = "La fecha de carga no puede ser futura")
